split danger into one-file-per-task
