+++ scripts/danger/task-greenkeeper.js
@@ -1,0 +1,7 @@
+// @flow
+
+import {message} from 'danger'
+
+export default function run() {
+  message('greenkeeper ran')
+}
