+++ scripts/danger/task-js-data.js
@@ -1,0 +1,36 @@
+// @flow
+
+import {readLogFile, fileLog, isBadDataValidationLog} from './lib'
+
+export default async function run() {
+  await data()
+  await busData()
+}
+
+function data() {
+  const dataValidationLog = readLogFile('./logs/validate-data')
+
+  if (!dataValidationLog) {
+    return
+  }
+
+  if (!isBadDataValidationLog(dataValidationLog)) {
+    return
+  }
+
+  fileLog("Something's up with the data.", dataValidationLog)
+}
+
+function busData() {
+  const busDataValidationLog = readLogFile('./logs/validate-bus-data')
+
+  if (!busDataValidationLog) {
+    return
+  }
+
+  if (!isBadDataValidationLog(busDataValidationLog)) {
+    return
+  }
+
+  fileLog("🚌 Something's up with the bus routes.", busDataValidationLog)
+}
