+++ scripts/danger/task-js-flow.js
@@ -1,0 +1,17 @@
+// @flow
+
+import {readLogFile, fileLog} from './lib'
+
+export default function run() {
+  const flowLog = readLogFile('./logs/flow')
+
+  if (!flowLog) {
+    return
+  }
+
+  if (flowLog === 'Found 0 errors') {
+    return
+  }
+
+  fileLog('Flow would like to interject about types…', flowLog)
+}
