+++ scripts/danger/task-js-prettier.js
@@ -1,0 +1,13 @@
+// @flow
+
+import {readLogFile, fileLog} from './lib'
+
+export default function run() {
+  const prettierLog = readLogFile('./logs/prettier')
+
+  if (!prettierLog) {
+    return
+  }
+
+  fileLog('Prettier made some changes', prettierLog, {lang: 'diff'})
+}
